Move all trakt functionality back to trakt-auth module
