+++ app/src/main/java/me/banes/chris/tivi/inject/AppComponent.kt
@@ -26,7 +26,6 @@ import me.banes.chris.tivi.jobs.JobsCreator
 import me.banes.chris.tivi.showdetails.ShowDetailsBuilder
 import me.banes.chris.tivi.tmdb.TmdbModule
 import me.banes.chris.tivi.trakt.TraktAuthModule
-import me.banes.chris.tivi.trakt.TraktModule
 import javax.inject.Singleton
 
 @Singleton
@@ -38,7 +37,6 @@ import javax.inject.Singleton
     ViewModelBuilder::class,
     HomeBuilder::class,
     ShowDetailsBuilder::class,
-    TraktModule::class,
     TraktAuthModule::class,
     TmdbModule::class,
     NetworkModule::class
